Fix for problem with object_class and OrderedHash
Changed parse_value to use the []= method to set the value for the key
rather than using rb_hash_aset when an object_class is specified.

rb_hash_aset doesn't allow the object_class to specify its own setter,
and this means that JSON.parse can't be used with something like
ActiveSupport::OrderedHash to support an ordered representation of the
parsed JSON.

To support this, a check for NIL_P(object_class) is done for each parse_value
call. It might be worth checking the benchmarks for the parser to see if this
has a noticeable effect.